hot fix

+++ app/Http/Controllers/SearchCompetitorsController.php
@@ -106,7 +106,7 @@ class SearchCompetitorsController extends Controller
     {
         $progress = CompetitorsProgressBar::where('page_hash', '=', $request->input('pageHash'))->first();
 
-        if ($progress->percent === 100) {
+        if (isset($progress->percent) && $progress->percent === 100) {
             $progress->delete();
             return response()->json([
                 'percent' => $progress->percent,
